Move resource.Type to its own file.
