+++ resource/resources_test.go
@@ -14,85 +14,6 @@ var _ = gc.Suite(&resourceSuite{})
 
 type resourceSuite struct{}
 
-func (s *resourceSuite) TestParseTypeOkay(c *gc.C) {
-	rt, ok := resource.ParseType("file")
-
-	c.Check(ok, jc.IsTrue)
-	c.Check(rt, gc.Equals, resource.TypeFile)
-}
-
-func (s *resourceSuite) TestParseTypeRecognized(c *gc.C) {
-	supported := []resource.Type{
-		resource.TypeFile,
-	}
-	for _, expected := range supported {
-		rt, ok := resource.ParseType(expected.String())
-
-		c.Check(ok, jc.IsTrue)
-		c.Check(rt, gc.Equals, expected)
-	}
-}
-
-func (s *resourceSuite) TestParseTypeEmpty(c *gc.C) {
-	rt, ok := resource.ParseType("")
-
-	c.Check(ok, jc.IsFalse)
-	c.Check(rt, gc.Equals, resource.TypeUnknown)
-}
-
-func (s *resourceSuite) TestParseTypeUnsupported(c *gc.C) {
-	rt, ok := resource.ParseType("spam")
-
-	c.Check(ok, jc.IsFalse)
-	c.Check(rt, gc.Equals, resource.Type("spam"))
-}
-
-func (s *resourceSuite) TestTypeStringSupported(c *gc.C) {
-	supported := map[resource.Type]string{
-		resource.TypeFile: "file",
-	}
-	for rt, expected := range supported {
-		str := rt.String()
-
-		c.Check(str, gc.Equals, expected)
-	}
-}
-
-func (s *resourceSuite) TestTypeStringUnknown(c *gc.C) {
-	str := resource.TypeUnknown.String()
-
-	c.Check(str, gc.Equals, "<unknown>")
-}
-
-func (s *resourceSuite) TestTypeStringUnsupported(c *gc.C) {
-	str := resource.Type("spam").String()
-
-	c.Check(str, gc.Equals, "spam")
-}
-
-func (s *resourceSuite) TestTypeValidateSupported(c *gc.C) {
-	supported := []resource.Type{
-		resource.TypeFile,
-	}
-	for _, rt := range supported {
-		err := rt.Validate()
-
-		c.Check(err, jc.ErrorIsNil)
-	}
-}
-
-func (s *resourceSuite) TestTypeValidateUnknown(c *gc.C) {
-	err := resource.TypeUnknown.Validate()
-
-	c.Check(err, gc.ErrorMatches, `unsupported resource type .*`)
-}
-
-func (s *resourceSuite) TestTypeValidateUnsupported(c *gc.C) {
-	err := resource.Type("spam").Validate()
-
-	c.Check(err, gc.ErrorMatches, `unsupported resource type .*`)
-}
-
 func (s *resourceSuite) TestParseOkay(c *gc.C) {
 	name := "my-resource"
 	data := map[string]interface{}{
